client/requirements.txt
pyzmq
schedule
requests
oauthlib==2.1.0
requests-oauthlib==1.1.0